+++ src/word_vector_dictionary.hpp
@@ -32,7 +32,7 @@ limitations under the License.
 
 namespace resembla {
 
-template<typename string_type, typename value_type = double, typename id_type = long>
+template<typename string_type, typename value_type = float, typename id_type = long>
 class WordVectorDictionary
 {
 public:
